adding explicit coding history kludge
There is an issue with AudioInspector and BatchInspector where either
program will crash if the size of the BEXT chunk is reported as uneven.

This kludge adds a padding space to coding history if this will force
the size of the BEXT chunk to be reported as even.

Adding as explicit class so it's easy to find, etc.