adding logic to handle sound field schema change + tests
if an xml field associated with a lookups dictionary  has no child
elements, the engine will attempt to resolve its value against a known
value in the dictionary (case-insensitive). If that value exists in the
dictionary, it will use it, otherwise, it will return an empty string.

for example: <sound_field>mono</sound_field> will return "Mono," if
"Mono" exists as a valid value in the lookups dictionary.